Changed display of selected skin in the store and fixed bugs.

+++ src/main/java/Entity/BallEnemy.java
@@ -10,7 +10,7 @@ import static Utilz.Constants.BallDetails.*;
 
 public class BallEnemy extends Entity {
 
-    public BallEnemy(Game game, int xPos, int yPos, float speed) {
+    public BallEnemy(Game game, int xPos, int yPos, int speed) {
         super(game, Type.bad_bullet, xPos, yPos, BALL_WIDTH, BALL_HEIGHT, speed, 1, RUNNING);
         name = "Ball";
         dir = Direction.fall;
