+++ src/main/java/Entity/Money.java
@@ -10,7 +10,7 @@ import static Utilz.Constants.MoneyDetails.*;
 
 public class Money extends Entity {
 
-    public Money(Game game, int xPos, int yPos, float speed) {
+    public Money(Game game, int xPos, int yPos, int speed) {
         super(game, Type.money, xPos, yPos, MONEY_WIDTH, MONEY_HEIGHT, speed, 1, RUNNING);
         name = "Money";
         dir = Direction.fall;
